+++ internal/helpers/shelper.go
@@ -117,3 +117,17 @@ func PickFromSemverStrings(list []string, contraint string) ([]string, error) {
 
 	return semvers_filtered, nil
 }
+
+func Constrained(version string, constraint string) (bool, error) {
+	c, err := semver.NewConstraint(constraint)
+	if err != nil {
+		return false, err
+	}
+
+	v, err := semver.NewVersion(version)
+	if err != nil {
+		return false, err
+	}
+
+	return c.Check(v), nil
+}