+++ test/main.tf
@@ -145,3 +145,33 @@ output "last_local" {
 output "last_local_noprerelease" {
   value = local.local_last_no_prerelease
 }
+
+output "semvers_constrained_results" {
+  value = [
+    {
+      expected  = true
+      arguments = "0.1.1, >= 0.1"
+      result    = provider::semvers::constrained("0.1.1", ">= 0.1")
+    },
+    {
+      expected  = false
+      arguments = "0.1, >= 0.1.1"
+      result    = provider::semvers::constrained("0.1", ">= 0.1.1")
+    },
+    {
+      expected  = true
+      arguments = "0.1.2, ~> 0.1"
+      result    = provider::semvers::constrained("0.1.2", "~> 0.1")
+    },
+    {
+      expected  = false
+      arguments = "0.2, ~> 0.1"
+      result    = provider::semvers::constrained("0.2", "~> 0.1")
+    },
+    {
+      expected  = true
+      arguments = "0.2, 0.1.1 - 0.2.0"
+      result    = provider::semvers::constrained("0.2", "0.1.1 - 0.2.0")
+    },
+  ]
+}